Move child-thread request context test into test_routes.py
